If metadata defines a box that isn't there then warn, but don't fail

+++ lib/bento/upload.rb
@@ -37,19 +37,23 @@ class UploadRunner
     md_data = box_metadata(md_file)
 
     md_data['providers'].each_pair do |prov, prov_data|
-      banner("Uploading bento/#{md_data['name']} version:#{md_data['version']} provider:#{prov}...")
+      if File.exist?(File.join('builds', prov_data['file']))
+        banner("Uploading bento/#{md_data['name']} version:#{md_data['version']} provider:#{prov}...")
 
-      upload_cmd = "vagrant cloud publish bento/#{md_data['name']} #{md_data['version']} #{prov} builds/#{prov_data['file']} --description '#{box_desc(md_data['name'])}' --short-description '#{box_desc(md_data['name'])}' --version-description '#{ver_desc(md_data, prov)}' --force --release"
-      shellout(upload_cmd)
+        upload_cmd = "vagrant cloud publish bento/#{md_data['name']} #{md_data['version']} #{prov} builds/#{prov_data['file']} --description '#{box_desc(md_data['name'])}' --short-description '#{box_desc(md_data['name'])}' --version-description '#{ver_desc(md_data, prov)}' --force --release"
+        shellout(upload_cmd)
 
-      slug_name = lookup_slug(md_data['name'])
-      next if slug_name.nil?
-      banner("Uploading slug bento/#{slug_name} from #{md_data['name']} version:#{md_data['version']} provider:#{prov}...")
-      upload_cmd = "vagrant cloud publish bento/#{slug_name} #{md_data['version']} #{prov} builds/#{prov_data['file']} --description '#{slug_desc(slug_name)}' --short-description '#{slug_desc(slug_name)}' --version-description '#{ver_desc(md_data, prov)}' --force --release"
-      shellout(upload_cmd)
+        slug_name = lookup_slug(md_data['name'])
+        next if slug_name.nil?
+        banner("Uploading slug bento/#{slug_name} from #{md_data['name']} version:#{md_data['version']} provider:#{prov}...")
+        upload_cmd = "vagrant cloud publish bento/#{slug_name} #{md_data['version']} #{prov} builds/#{prov_data['file']} --description '#{slug_desc(slug_name)}' --short-description '#{slug_desc(slug_name)}' --version-description '#{ver_desc(md_data, prov)}' --force --release"
+        shellout(upload_cmd)
 
-      # move the box file to the completed directory
-      FileUtils.mv(File.join('builds', prov_data['file']), File.join('builds', 'uploaded', prov_data['file']))
+        # move the box file to the completed directory
+        FileUtils.mv(File.join('builds', prov_data['file']), File.join('builds', 'uploaded', prov_data['file']))
+      else # box in metadata isn't on disk
+        warn "The #{prov} box defined in the metadata file #{md_file} does not exist at builds/#{prov_data['file']}. Skipping!"
+      end
     end
 
     # move the metadata file to the completed directory
